Создание dto классов
