+++ app/src/main/java/ru/practicum/android/diploma/filters/data/dto/FilterCitiesDto.kt
@@ -1,0 +1,7 @@
+package ru.practicum.android.diploma.filters.data.dto
+
+data class FilterCitiesDto(
+    //тут есть еще дочерние регионы
+    val id: String,
+    val name: String,
+)
