+++ app/src/main/java/ru/practicum/android/diploma/filters/data/dto/FilterCitiesRequest.kt
@@ -1,0 +1,4 @@
+package ru.practicum.android.diploma.filters.data.dto
+
+class FilterCitiesRequest(val expression: String)
+    // https://api.hh.ru/areas/{area_id}
