+++ app/src/main/java/ru/practicum/android/diploma/filters/data/dto/FilterCountryRequest.kt
@@ -1,0 +1,4 @@
+package ru.practicum.android.diploma.filters.data.dto
+
+class FilterCountryRequest(val expression: String)
+    // https://api.hh.ru/areas/countries
