+++ app/src/main/java/ru/practicum/android/diploma/filters/data/dto/FilterCountryResponse.kt
@@ -1,0 +1,8 @@
+package ru.practicum.android.diploma.filters.data.dto
+
+import ru.practicum.android.diploma.util.network.Response
+
+class FilterCountryResponse(
+    val id: String,
+    val name: String,
+) : Response()
