+++ app/src/main/java/ru/practicum/android/diploma/filters/data/dto/FilterIndustriesRequest.kt
@@ -1,0 +1,4 @@
+package ru.practicum.android.diploma.filters.data.dto
+
+class FilterIndustriesRequest(val expression: String)
+    // https://api.hh.ru/industries
