+++ app/src/main/java/ru/practicum/android/diploma/search/data/dto/VacancySearchRequest.kt
@@ -1,3 +1,3 @@
-package ru.practicum.android.diploma.search.data
+package ru.practicum.android.diploma.search.data.dto
 
 data class VacancySearchRequest(val expression: String)
