+++ app/src/main/java/ru/practicum/android/diploma/search/data/network/HHApiService.kt
@@ -2,7 +2,7 @@ package ru.practicum.android.diploma.search.data.network
 
 import retrofit2.http.GET
 import retrofit2.http.Query
-import ru.practicum.android.diploma.search.data.VacancySearchResponse
+import ru.practicum.android.diploma.search.data.dto.VacancySearchResponse
 
 interface HHApiService {
     @GET("vacancies?text={expression}")
