+++ app/src/main/java/ru/practicum/android/diploma/util/dto/KeySkillDto.kt
@@ -1,0 +1,5 @@
+package ru.practicum.android.diploma.util.dto
+
+data class KeySkillDto(
+    val name: String
+)
