+++ app/src/main/java/ru/practicum/android/diploma/util/dto/SalaryDto.kt
@@ -1,4 +1,4 @@
-package ru.practicum.android.diploma.vacancy.data.dto
+package ru.practicum.android.diploma.util.dto
 
 data class SalaryDto(
     val currency: String?,
