+++ app/src/main/java/ru/practicum/android/diploma/util/dto/ScheduleDto.kt
@@ -1,4 +1,4 @@
-package ru.practicum.android.diploma.vacancy.data.dto
+package ru.practicum.android.diploma.util.dto
 
 data class ScheduleDto(
     val id: String?,
