+++ app/src/main/java/ru/practicum/android/diploma/vacancy/data/dto/VacancyDetailsRequest.kt
@@ -1,5 +1,3 @@
 package ru.practicum.android.diploma.vacancy.data.dto
 
-data class KeySkillDto(
-    val name: String
-)
+class VacancyDetailsRequest(val expression: String)
