--- app/src/main/java/ru/practicum/android/diploma/vacancy/data/dto/VacancyDto.kt
@@ -1,15 +1,0 @@
-package ru.practicum.android.diploma.vacancy.data.dto
-
-data class VacancyDto(
-    val id: String,
-    val name: String,
-    val salary: SalaryDto?,
-    val employer: EmployerDto?,
-    val area: AreaDto,
-    val address: AddressDto?,
-    val experience: ExperienceDto?,
-    val schedule: ScheduleDto?,
-    val employment: EmploymentDto?,
-    val description: String,
-    val keySkills: List<KeySkillDto>,
-)
